moved home.php to dynamic folder, renamed pages_controller to dynamic_controller, added dynamic controller access to routes.php, and uncommented some script to allow lists and calendar to work in search_events.php
