+++ controllers/dynamic_controller.php
@@ -1,7 +1,7 @@
 <?php
-  class PagesController {
+  class DynamicController {
     public function home() {
-      require_once('views/pages/home.php');
+      require_once('views/dynamic/home.php');
     }
 
     public function error() {
