+++ views/templates/header.php
@@ -1,6 +1,6 @@
 <div class="container-fluid">
     <nav class="navbar sticky-top navbar-expand-sm bg-light justify-content-end">
-        <a class="navbar-brand" href="?controller=static&action=landing">Kamps For Kids</a>
+        <a class="navbar-brand" href="?controller=static&action=landing">Kamparoos</a>
         <ul class="navbar-nav">
             <li class="nav-item">
                 <a class="nav-link" href="?controller=static&action=registration">Sign Up</a>
